+++ rlw/upgrade.ps1
@@ -1,3 +1,17 @@
+# ---
+# RightScript Name: RL10 Windows Upgrade
+# Description: Check whether a RightLink upgrade is available and perform the upgrade.
+# Inputs:
+#   UPGRADES_FILE_LOCATION:
+#     Input Type: single
+#     Category: RightScale
+#     Description: External location of 'upgrades' file
+#     Default: text:https://rightlink.rightscale.com/rightlink/upgrades
+#     Required: false
+#     Advanced: true
+# ...
+#
+
 # Will compare current version of rightlink 'running' with latest version provided from 'upgrades'
 # file. If they differ, will update to latest version.  Note that latest version can be an older version
 # if a downgrade is best.