+++ rlw/wait-for-eip.ps1
@@ -1,5 +1,10 @@
-# On AWS, if an elastic IP address is specified this script will wait until all API Hosts are
-# reporting the expected IP address for the server, which is recorded in the user-data
+# ---
+# RightScript Name: RL10 Linux Wait For EIP
+# Description: On AWS, if an elastic IP address is specified this script will
+#   wait until all API Hosts are reporting the expected IP address for the server,
+#   which is recorded in the user-data
+# ...
+#
 
 $RS_DIR = 'C:\ProgramData\RightScale\RightLink'
 $RS_ID_FILE = "${RS_DIR}\rightscale-identity"